requirements.txt

viam-sdk==0.41.1

typing-extensions
hx711==1.1.2.3
RPi.GPIO